+++ config/paymaya-sdk.php
@@ -19,5 +19,19 @@ return [
         Webhook::CHECKOUT_SUCCESS => 'api/payment-callback/paymaya/success',
         Webhook::CHECKOUT_FAILURE => 'api/payment-callback/paymaya/failure',
         Webhook::CHECKOUT_DROPOUT => 'api/payment-callback/paymaya/dropout',
+
+//        Webhook::PAYMENT_SUCCESS => 'api/test/success',
+//        Webhook::PAYMENT_EXPIRED => 'api/test/expired',
+//        Webhook::PAYMENT_FAILED => 'api/test/failed',
+    ],
+
+    'checkout' => [
+        'customization' => [
+            'logoUrl' => 'http://testimge1',
+            'iconUrl' => 'http://testimge1',
+            'appleTouchIconUrl' => 'http://testimge1',
+            'customTitle' => 'test paymaya sandbox title',
+            'colorScheme' => '#e01c44',
+        ],
     ],
 ];