+++ src/LaravelPaymaya.php
@@ -4,8 +4,10 @@ namespace Lloricode\LaravelPaymaya;
 
 use Lloricode\LaravelPaymaya\Facades\CheckoutFacade;
 use Lloricode\LaravelPaymaya\Facades\ClientFacade;
+use Lloricode\LaravelPaymaya\Facades\CustomizationFacade;
 use Lloricode\LaravelPaymaya\Facades\WebhookFacade;
 use Lloricode\Paymaya\Client\Checkout\CheckoutClient;
+use Lloricode\Paymaya\Client\Checkout\CustomizationClient;
 use Lloricode\Paymaya\Client\WebhookClient;
 use Lloricode\Paymaya\PaymayaClient;
 
@@ -24,5 +26,10 @@ class LaravelPaymaya
     public static function webhook(): WebhookClient
     {
         return WebhookFacade::getFacadeRoot();
+    }
+
+    public static function customization(): CustomizationClient
+    {
+        return CustomizationFacade::getFacadeRoot();
     }
 }