+++ tests/Commands/WebhookCommandTest.php
@@ -29,17 +29,16 @@ class WebhookCommandTest extends TestCase
         PaymayaFacade::client()->setHandlerStack($handlerStack);
 
         $this->artisan('paymaya-sdk:webhook:retrieve')
-//            ->expectsTable(
-//                [
-//                    'id',
-//                    'name',
-//                    'callbackUrl',
-//                    'createdAt',
-//                    'updatedAt',
-//                ],
-//                [self::sampleWebhookData()]
-//            )
-            // TODO: commented because its always failed
+            ->expectsTable(
+                [
+                    'id',
+                    'name',
+                    'callbackUrl',
+                    'createdAt',
+                    'updatedAt',
+                ],
+                [self::sampleWebhookData()]
+            )
             ->assertExitCode(0);
     }